flatcar: extract filesystem contents from disk image
instead of resizing. On cloud instances with limited block IO throughput, the
current scheme of extending the disk image with 3000MB of zeros and resizing
the filesystem is slow. Instead, extract the partition contents with tar, and
dispose of the disk image after use. This makes cleanup and other operations
faster.