flatcar: use correct libc/ld to run saved binutils
Ubuntu 18.04 has a different (older) version of libc than the one required by
Flatcar's binutils. To ensure that binutils can be executed correctly when
relinking outside of the flatcar chroot, save the libraries and launch by
manually specifying the correct dynamic linker. The LD_LIBRARY_PATH definition
needs to be moved so that the saved libc is only used for launching ld, and not
the other binaries (grep/cp) that run using ubuntu's dynlinker.